CenterPane: extract LineItemForm and FileView
